scripts: imx-usb-loader: Use dcd len to invalidate dcd data
We invalidate the dcd data in the uploaded image since we already
processed it manually. To do so we have set the dcd pointer to 0.
Doing it this way prevents the ROM from executing the HAB code in
debug mode. Use the dcd length instead to invalidate the dcd data.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>