mtd: mtdoob device: change name to have the chip name first
Normally all device files assoctiated to a mtd device begin with
the chip name itself. Only the mtdoob device is an exception:
it begins with the chip name without the index end ends with the
index. Change the name to be like the other names, i.e. change
nand_oob0 to nand0.oob.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>