ds1307: Configure ds1341 for lowest power mode
Empirical observations show that configuring INTCN=1, ECLK=0, EGFIL=0,
DOSF=1 on DS1341 put the chip in the mode where it draws the least
amount fo current.

Add code to configure DS1341 in such a way in case Barebox is the last
code that runs on the processor before being shut down.

Signed-off-by: Andrey Smirnov <andrew.smirnov@gmail.com>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>