net: usb: use minimum timeout when polling for new packets
When no new packets have arrived we want to return to the caller
as soon as possible to give other network controllers the chance
to receive packets. With the current USB bulk message timeout of
one second other network controllers do not work properly whenever
the USB network controller is active.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>